First communication with the iRobot Create.
I had to bring in the source code for the usb/serial drivers. For
some reason, the way it was packaged didn't work with the
application.

I was able to set up permissions to use the device, open it and
send some initialization commands to it.

Now we will start the refactoring process.