Working version.
Upon initial end-to-end testing, we found a couple of defects. In
particular, when a bad checksum is found, we now clear the packet; we
were getting a buffer overflow if we didn't do that. Another was
adjusting the sleep time for the USB read down to 1/10 second. That
seems to be a happy medium.

Finally, all the extraneous test code in the mediation activity was
deleted and the view was cleaned up with some explanatory text.